+++ ratelimiter/ratelimiter.go
@@ -46,8 +46,56 @@ type Client interface {
 	) (err error)
 }
 
-func WithTimeout(timeout time.Duration) options.AcquireOption {
-	return options.WithTimeout(timeout)
+// WithDecrease specifies decrease timeout from context
+// Default value defined in options.DecreaseTimeout
+// Example 1:
+//   ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+//   defer cancel()
+//   err := db.Ratelimiter().AcquireResource(
+//     ctx,
+//     "/path",
+//     "resource",
+//     1000,
+//     ratelimiter.WithAcquire(),
+//     ratelimiter.WithDecrease(100 * time.Millisecond),
+//   )
+// Request to ratelimiter service will have:
+//   - context with timeout 1000ms
+//   - cancelAfter timeout (1000-100)ms
+//   - operation timeout - from main config
+//
+// Example 2:
+//   ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+//   defer cancel()
+//   err := db.Ratelimiter().AcquireResource(
+//     ctx,
+//     "/path",
+//     "resource",
+//     1000,
+//     ratelimiter.WithReportSync(),
+//     ratelimiter.WithDecrease(100 * time.Millisecond),
+//   )
+// Request to ratelimiter service will have:
+//   - context with timeout 1000ms
+//   - operation timeout (1000-100)ms
+//   - cancelAfter timeout - from main config
+//
+// Example 3:
+//   ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+//   defer cancel()
+//   err := db.Ratelimiter().AcquireResource(
+//     ctx,
+//     "/path",
+//     "resource",
+//     1000,
+//     ratelimiter.WithReportAsync(),
+//     ratelimiter.WithDecrease(100 * time.Millisecond), // nop
+//   )
+// Request to ratelimiter service will have:
+//   - context with timeout 1000ms
+//   - operation timeout and cancelAfter timeout - from main config
+func WithDecrease(decreaseTimeout time.Duration) options.AcquireOption {
+	return options.WithDecrease(decreaseTimeout)
 }
 
 func WithAcquire() options.AcquireOption {